requirements.txt
about-time==4.2.1
alive-progress==3.1.5
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autograd==1.6.2
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cma==3.2.2
colorama==0.4.6
colorlog==6.8.0
comm==0.2.0
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.9.0
Flask==3.0.1
Flask-Cors==4.0.0
fonttools==4.46.0
fqdn==1.5.1
frozendict==2.4.0
fsspec==2023.4.0
future==0.18.3
grapheme==0.6.0
greenlet==3.0.3
html5lib==1.1
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
install==1.3.5
ipykernel==6.27.1
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
minio==7.2.3
mistune==3.0.2
mplfinance==0.12.10b0
mpmath==1.3.0
multitasking==0.0.11
nacos-sdk-python==0.1.14
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.3.1
parso==0.8.3
peewee==3.17.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.6
pure-eval==0.2.2
pyaml==23.12.0
pycparser==2.21
pycryptodome==3.20.0
pyface==8.0.0
Pygments==2.17.2
pymoo==0.6.1.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.1
pyswarm==0.6
pyswarms==1.3.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scikit-optimize==0.9.0
scipy==1.10.1
seaborn==0.13.1
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tornado==6.4
tqdm==4.64.1
traitlets==5.14.0
traits==6.4.3
traitsui==8.0.0
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.37.1
widgetsnbextension==4.0.9
wrapt==1.16.0
xgboost==2.0.3
yfinance==0.2.35
zipp==3.17.0
